requirements.txt
google-cloud-logging==3.2.4
Flask==2.2.2
functions-framework==3.2.0